Refactor death effects logic and update character relationship handling
